init: decouple ECC init from primitives
No behavior change.

One entry point for ECC initialization, which will soon be moved to
the consensus library, and set up a verification context.

The goal is to decouple context handling from the primitive key.h/cpp
and pubkey.h/cpp, which shouldn't be accessed by the upper layers.
