+++ src/CMakeLists.txt
@@ -103,6 +103,7 @@ add_library(bitcoin_common STATIC EXCLUDE_FROM_ALL
   common/args.cpp
   common/bloom.cpp
   common/config.cpp
+  common/ecc_init.cpp
   common/init.cpp
   common/interfaces.cpp
   common/license_info.cpp
