+++ src/bench/bip324_ecdh.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <bench/bench.h>
+#include <common/ecc_init.h>
 #include <key.h>
 #include <pubkey.h>
 #include <random.h>
@@ -15,7 +16,7 @@
 
 static void BIP324_ECDH(benchmark::Bench& bench)
 {
-    ECC_Context ecc_context{};
+    const auto ecc_context{MakeContextECC()};
     FastRandomContext rng;
 
     std::array<std::byte, 32> key_data;
