+++ src/bench/verify_script.cpp
@@ -5,6 +5,7 @@
 #include <addresstype.h>
 #include <bench/bench.h>
 #include <coins.h>
+#include <common/ecc_init.h>
 #include <key.h>
 #include <policy/policy.h>
 #include <primitives/transaction.h>
@@ -45,7 +46,7 @@ static size_t ExpectedWitnessStackSize(ScriptType script_type)
 // Microbenchmark for verification of standard scripts.
 static void VerifyScriptBench(benchmark::Bench& bench, ScriptType script_type)
 {
-    ECC_Context ecc_context{};
+    const auto ecc_context{MakeContextECC()};
 
     // Create deterministic key material needed for output script creation / signing
     CKey privkey;
