+++ src/bitcoin-wallet.cpp
@@ -8,11 +8,11 @@
 #include <chainparamsbase.h>
 #include <clientversion.h>
 #include <common/args.h>
+#include <common/ecc_init.h>
 #include <common/license_info.h>
 #include <common/system.h>
 #include <compat/compat.h>
 #include <interfaces/init.h>
-#include <key.h>
 #include <logging.h>
 #include <pubkey.h>
 #include <tinyformat.h>
@@ -124,7 +124,7 @@ MAIN_FUNCTION
         return EXIT_FAILURE;
     }
 
-    ECC_Context ecc_context{};
+    const auto ecc_context{MakeContextECC()};
     if (!wallet::WalletTool::ExecuteWalletToolFunc(args, command->command)) {
         return EXIT_FAILURE;
     }
