+++ src/node/context.cpp
@@ -9,7 +9,6 @@
 #include <interfaces/chain.h>
 #include <interfaces/mining.h>
 #include <kernel/context.h>
-#include <key.h>
 #include <net.h>
 #include <net_processing.h>
 #include <netgroup.h>
