+++ src/node/interfaces.cpp
@@ -10,6 +10,7 @@
 #include <chainparams.h>
 #include <coins.h>
 #include <common/args.h>
+#include <common/ecc_init.h>
 #include <common/settings.h>
 #include <consensus/amount.h>
 #include <consensus/merkle.h>
@@ -25,7 +26,6 @@
 #include <interfaces/rpc.h>
 #include <interfaces/types.h>
 #include <kernel/context.h>
-#include <key.h>
 #include <logging.h>
 #include <mapport.h>
 #include <net.h>
@@ -133,7 +133,7 @@ public:
 
         m_context->warnings = std::make_unique<node::Warnings>();
         m_context->kernel = std::make_unique<kernel::Context>();
-        m_context->ecc_context = std::make_unique<ECC_Context>();
+        m_context->ecc_context = MakeContextECC();
         if (!AppInitSanityChecks(*m_context->kernel)) return false;
 
         if (!AppInitLockDirectories()) return false;
