+++ src/test/fuzz/key.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <chainparams.h>
+#include <common/ecc_init.h>
 #include <key.h>
 #include <key_io.h>
 #include <outputtype.h>
@@ -33,7 +34,7 @@
 
 void initialize_key()
 {
-    static ECC_Context ecc_context{};
+    static const auto ecc_context{MakeContextECC()};
     SelectParams(ChainType::REGTEST);
 }
