+++ src/test/fuzz/message.cpp
@@ -3,6 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include <chainparams.h>
+#include <common/ecc_init.h>
 #include <common/signmessage.h>
 #include <key_io.h>
 #include <test/fuzz/FuzzedDataProvider.h>
@@ -19,7 +20,7 @@
 
 void initialize_message()
 {
-    static ECC_Context ecc_context{};
+    static const auto ecc_context{MakeContextECC()};
     SelectParams(ChainType::REGTEST);
 }
