+++ src/test/fuzz/secp256k1_ecdsa_signature_parse_der_lax.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <key.h>
+#include <common/ecc_init.h>
 #include <secp256k1.h>
 #include <test/fuzz/FuzzedDataProvider.h>
 #include <test/fuzz/fuzz.h>
@@ -26,7 +26,7 @@ FUZZ_TARGET(secp256k1_ecdsa_signature_parse_der_lax)
     secp256k1_ecdsa_signature sig_der_lax;
     const bool parsed_der_lax = ecdsa_signature_parse_der_lax(&sig_der_lax, signature_bytes.data(), signature_bytes.size()) == 1;
     if (parsed_der_lax) {
-        ECC_Context ecc_context{};
+        const auto ecc_context{MakeContextECC()};
         (void)SigHasLowR(&sig_der_lax);
     }
 }
