+++ src/test/sanity_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <key.h>
+#include <common/ecc_init.h>
 #include <test/util/setup_common.h>
 
 #include <boost/test/unit_test.hpp>
