+++ src/test/util/setup_common.cpp
@@ -9,6 +9,7 @@
 #include <chain.h>
 #include <chainparams.h>
 #include <coins.h>
+#include <common/ecc_init.h>
 #include <common/system.h>
 #include <consensus/amount.h>
 #include <consensus/consensus.h>
@@ -237,7 +238,7 @@ BasicTestingSetup::BasicTestingSetup(const ChainType chainType, TestOpts opts)
     LogInstance().StartLogging();
     m_node.warnings = std::make_unique<node::Warnings>();
     m_node.kernel = std::make_unique<kernel::Context>();
-    m_node.ecc_context = std::make_unique<ECC_Context>();
+    m_node.ecc_context = MakeContextECC();
     SetupEnvironment();
 
     m_node.chain = interfaces::MakeChain(m_node);
